+++ Cursus/ft_printf/ft_printf.h
@@ -6,7 +6,7 @@
 /*   By: mmauchre <mmauchre@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/17 12:26:10 by mmauchre          #+#    #+#             */
-/*   Updated: 2023/12/12 19:50:21 by mmauchre         ###   ########.fr       */
+/*   Updated: 2023/12/13 21:33:05 by mmauchre         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -31,8 +31,8 @@ void	ft_flag_p(unsigned long int n, int *total_return);
 int		check_bonus(char c);
 int		ft_isdigit(int c);
 void	ft_putstr(char *str, int *j, int *total_return);
-void	ft_flag_dieze_and_x(unsigned long int n, int *total_return);
-void	ft_flag_dieze_and_bigx(unsigned long int n, int *total_return);
+void	ft_flag_dieze_and_x(unsigned int n, int *total_return);
+void	ft_flag_dieze_and_bigx(unsigned int n, int *total_return);
 void	ft_flag_di_and_space(int n, int *total_return);
 void	ft_flag_di_and_plus(int n, int *total_return);
